created createPin controller
